Reorganize commit-mgr/blockchain files
